refactor: rename PromptLedgerConnectionPopup to ConnectLedgerPopup
